relocate metadata encryption - part 3
